Agents/master-agent/requirements.txt
fastapi
uvicorn
openai
google-genai
python-dotenv
pydantic
